Kalorien sind nun Int Darkmode gefixed leichte Versionsupgrades unter der Haube

+++ app/src/main/java/com/fourquestionmarks/food_o_mat/model/Meal.kt
@@ -14,7 +14,7 @@ data class Meal(
     @NotNull
     var category: String,
     @NotNull
-    var calories: Float,
+    var calories: Int,
     @NotNull
     var carbohydrates: Float,
     @NotNull
